pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "code-review-agent"
version = "0.1.0"
description = "Lightweight automated code review agent"
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["pyyaml>=6.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
